Protect against giant mirror directories.
This was a problem invoked by having large catalogue files
in imaginglss script directory.

tar-dir will issue an error and quite if there is more than 50 MB of
code to be mirrored. This should be a reasonable upper limit of size for
on an agile, developing code base that doesn't work via pip.

+++ activate.sh
@@ -40,7 +40,14 @@ if [[ -n $BCASTROOT ]]; then
     }
 
     function mirror {
-        local TMPFILE=`$DIRNAME/tar-dir.sh $*`
+        # BASH gimmicks: local always return 0
+        # http://unix.stackexchange.com/a/146900
+        local TMPFILE
+        TMPFILE=`$DIRNAME/tar-dir.sh $*`
+        if ! [ $? -eq 0 ]; then
+            # tar-dir must have failed
+            exit 1
+        fi
         bcast $TMPFILE
         rm $TMPFILE
     }
